Data type fix

+++ homework_01/main.py
@@ -23,7 +23,7 @@ def is_prime(num):
        функция для проверки входящих чисел на простое число
     """
     if num > 1:
-        for i in range(2,(num/2+1)):
+        for i in range(2,int(num/2+1)):
             if num % i == 0:
                 return False
             else:
